created forums.py
