+++ app/routes.py
@@ -6,30 +6,7 @@ from .models import User
 
 main = Blueprint('main', __name__)
 
-#dummy variable list of dictionaries 
-posts = [
-        {
-            'author': 'Sergey Astvatsaturov', 
-            'title': 'Single life!',
-            'content': 'I rock', 
-            'date_posted': 'February 10th, 2022'
-        },
 
-        {
-            'author': 'Stuart Anderson', 
-            'title': 'Life in Japan',
-            'content': "Can't wait to get back to the states", 
-            'date_posted': 'February 11th, 2022'
-        },
-          {
-            'author': 'Dante Samuels', 
-            'title': 'Kid 2',
-            'content': 'So excited!', 
-            'date_posted': 'February 13th, 2022'
-        }
-
-
-]
 
 #Routes are what you type into your browser to go to different webpages
 #use route decorators 
@@ -44,6 +21,7 @@ def home():
     return render_template('home.html')
 
 
+
 #create about page
 @main.route("/about")
 def about():
@@ -53,10 +31,8 @@ def about():
 @main.route("/products")
 def products():
     return render_template('products.html', title ='products')
+         
 
-@main.route("/forums")
-def forums():
-    return render_template('forums.html', title ='forums', posts = posts)
 
 @main.route('/profile')
 @login_required
